addressing issues (hopefully)
i've noticed that there's potentially an issue with the program in that it fails to account for the small period of time in between switching songs to where if it takes too long to check for the song (e.g. switching between pages and clicking on songs), it'll return all values for artist, song, and URI as none values, and that's not good. it's still a bit faulty, but it still works with queues, playlists, and albums, so long as you're not manually switching through all the songs all the time. i'm working on fixing this - but it might just be an issue with my computer as well. not entirely sure what the cause even is either, but this is my best guess.